backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.0
python-dotenv==1.0.1
httpx==0.27.2
psutil
requests
customtkinter
google-generativeai
google-genai
# ML Dependencies for Smart Brain
sentence-transformers
scikit-learn
aiohttp
